Various updates
Router program:
- Support for USB bluetooth adapters (rfcomm)
- Support for non-point-to-point interfaces
- Can manually set sync rates

Net Monitor:
- Now shows speeds in MBps with 1 decimal place (0.0 - 99.9MBps)

+++ router/src/typedefs.h
@@ -10,12 +10,6 @@
 #define TYPEDEFS_H_
 
 #include <stdbool.h>
-#include <stdint.h>
-#include <string.h>
-#include <curl/curl.h>
-#include <regex.h>
-#include <pthread.h>
-#include <time.h>
 
 #define UNUSED(x) (void)(x)
 
@@ -23,25 +17,26 @@ typedef uint8_t byte;
 typedef unsigned int uint;
 typedef unsigned long ulong;
 
+// cURL Received data from web request
 typedef struct {
 	char* memory;
 	size_t size;
 }s_memoryChunk;
 
 typedef struct {
-	CURL* c;
-	time_t lastUpdate;
-	int interval;
-	pthread_t hThread;
-	bool threadBusy;
-	regex_t reg;
-	void* data;
-	void (*dataReadyCallback)(s_memoryChunk*);
+	CURL* c;			// cURL object for web request
+	time_t lastUpdate;	// Time of last update
+	int interval;		// How often to update (seconds)
+	pthread_t hThread;	// The thread doing the update
+	bool threadBusy;	// Update thread is running
+	regex_t reg;		// Regular expression for parsing the received data
+	void* data;			// Pointer for storing the parsed data
+	void (*dataReadyCallback)(s_memoryChunk*);	// Received data for parsing
 }s_updateData;
 
 typedef struct {
-	uint16_t up;
-	uint16_t down;
+	uint32_t up;
+	uint32_t down;
 }s_netUpDown;
 
 #endif /* TYPEDEFS_H_ */
